Move makeCtx into existing src/test-utils.ts
Drop the just-introduced strategy-game-factory/test-helpers folder: the
project already has src/test-utils.ts (home of makeEvents) for test-only
helpers, so makeCtx belongs there too.
