+++ src/components/games/discs-flip-or-remove/optimality.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { mapValues } from 'lodash';
 import { dummyEvents, type GameMoves } from '../../strategy-game-factory';
-import { makeCtx } from '../../strategy-game-factory/test-helpers';
+import { makeCtx } from '../../../test-utils';
 import { moves as gameMoves, smartBotStrategy } from './discs-flip-or-remove';
 
 // Exhaustive optimality check for both variants (max 6 and max 10 discs).
