+++ src/components/games/number-pyramid/number-pyramid.spec.ts
@@ -1,7 +1,6 @@
 import { moves } from './number-pyramid';
 import type { Board, Slot } from './strategy';
-import { makeCtx } from '../../strategy-game-factory/test-helpers';
-import { makeEvents } from '../../../test-utils';
+import { makeCtx, makeEvents } from '../../../test-utils';
 
 const active = (value: number): Slot => ({ value, state: 'active' });
