+++ src/components/games/totem-poles/triangular-grid-ropes-15/bot-strategy.spec.ts
@@ -2,7 +2,7 @@ import { sample } from 'lodash';
 import { getAllowedMoves, getAllowedSuperset, type Board, type Edge } from './helpers';
 import { findWinningMove } from './solver';
 import { smartBotStrategy } from './bot-strategy';
-import { makeCtx } from '../../../strategy-game-factory/test-helpers';
+import { makeCtx } from '../../../../test-utils';
 
 // Play a full game. Player 0 is the opponent, player 1 is the smart bot (the
 // second player, who has the winning strategy). Returns the index of the player
