+++ src/components/strategy-game-factory/game-parts/common/cta-text.spec.ts
@@ -1,5 +1,5 @@
 import { getCtaText } from './cta-text';
-import { makeCtx } from '../../test-helpers/make-ctx';
+import { makeCtx } from '../../../../test-utils';
 
 describe('getCtaText', () => {
   describe('vsHuman mode', () => {
