+++ src/components/strategy-game-factory/game-parts/game-sidebar/game-sidebar.spec.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router';
 import { GameSidebar, type SidebarMoves } from './game-sidebar';
 import type { Ctx } from '../../types';
-import { makeCtx } from '../../test-helpers/make-ctx';
+import { makeCtx } from '../../../../test-utils';
 
 beforeAll(() => {
   const { unmount } = renderSidebar();
