--- src/components/strategy-game-factory/test-helpers/index.ts
@@ -1,1 +1,0 @@
-export { makeCtx } from './make-ctx';
