--- src/components/strategy-game-factory/test-helpers/make-ctx.ts
@@ -1,14 +1,0 @@
-import type { Ctx } from '../types';
-
-export const makeCtx = (overrides: Partial<Ctx> = {}): Ctx => ({
-  phase: 'roleSelection',
-  isHumanVsHumanGame: false,
-  resolvedPlayerNames: ['Player 1', 'Player 2'],
-  currentPlayer: null,
-  isClientMoveAllowed: false,
-  winnerIndex: null,
-  chosenRoleIndex: null,
-  turnState: null,
-  moveCount: 0,
-  ...overrides
-});
